Using seperate entry for standalone flow
