+++ src/runtime/generateTPAParams.ts
@@ -1,9 +1,8 @@
 import {pickBy} from './utils/utils';
 import {wixStylesFontUtils} from './utils/wixStyleFontUtils';
 import {wixStylesColorUtils} from './utils/wixStylesColorUtils';
-import {IOptions} from './main';
 import {IS_RTL_PARAM} from './constants';
-import {ISiteColor, ISiteTextPreset, IStyleFont, IStyleParams} from './types';
+import {IOptions, ISiteColor, ISiteTextPreset, IStyleFont, IStyleParams} from './types';
 
 export interface ITPAParams {
   colors: {[index: string]: {value: string}};
