+++ standalone.js
@@ -1,0 +1,1 @@
+export {getProcessedCssWithConfig, getStaticCssWithConfig} from './dist/runtime/standalone';
